Issue #7895:extra code for RequireThis
