+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/coding/requirethis/InputRequireThisAnonymousEmpty.java
@@ -29,9 +29,11 @@ public class InputRequireThisAnonymousEmpty {
             }
 
             public int doSideEffect() {
-                return bar; // violation
+                return bar;
             }
         };
+
+        this.bar = 0;
 
         new AnonWithEmpty() {
             int anonMember = 0;
@@ -56,4 +58,17 @@ public class InputRequireThisAnonymousEmpty {
             }
         };
     }
+
+    public void anotherMethod() {
+
+        this.anonMethod(new AnonWithEmpty() {
+            private int member;
+
+            public void fooEmpty() {
+                this.member++;
+            }
+        });
+    }
+
+    public void anonMethod(AnonWithEmpty aw) { }
 }
