+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/coding/requirethis/InputRequireThisEnumInnerClassesAndBugs3.java
@@ -128,7 +128,7 @@ class Issue22403 {
     class Nested {
         void bar() {
             instanceMethod();
-            i++; // violation
+            i++;
         }
     }
 }
